refactor: move Int.ifNotZero to :common
This is used by ChangeManager, which is to be moved to :anki-common

Part of 20737
