+++ common/src/main/java/com/ichi2/anki/common/utils/ext/Int.kt
@@ -1,0 +1,8 @@
+// SPDX-License-Identifier: GPL-3.0-or-later
+
+package com.ichi2.anki.common.utils.ext
+
+inline fun Int.ifNotZero(block: (value: Int) -> Unit) {
+    if (this == 0) return
+    block(this)
+}
